// src/lib/api/ftcevents-types.ts
export interface FtcScheduledMatch {
  description: string | null;
  field: string | null;
  tournamentLevel: string | null;
  startTime: string | null;
  series: number;
  matchNumber: number;
  teams: FtcScheduledMatchTeam[] | null;
  modifiedOn: string | null;
}

/** Scheduled Match Team */
export interface FtcScheduledMatchTeam {
  teamNumber: number | null;
  displayTeamNumber: string | null;
  station: string | null;
  team: string | null;
  teamName: string | null;
  surrogate: boolean;
  noShow: boolean;
}

/** Match Result Information */
export interface FtcMatchResult {
  actualStartTime: string | null;
  description: string | null;
  tournamentLevel: string | null;
  series: number;
  matchNumber: number;
  scoreRedFinal: number;
  scoreRedFoul: number;
  scoreRedAuto: number;
  scoreBlueFinal: number;
  scoreBlueFoul: number;
  scoreBlueAuto: number;
  postResultTime: string | null;
  teams: FtcMatchResultTeam[] | null;
  modifiedOn: string | null;
}

/** Match Result Team */
export interface FtcMatchResultTeam {
  teamNumber: number;
  station: string | null;
  dq: boolean;
  onField: boolean;
}